feat: remove books from library via butotn
I added the ability to remove books from the library. The removeBook function removes a book from the myLibrary array based on the provided index and updates the display. I also implemented the addRemoveButton function, which adds a Remove button to each book card and associates it with the corresponding book's index using the data-index attribute. When the button is clicked, it triggers the removeBook function with the correct index obtained from the data attribute.