refactor: library display to use a table
After reconsidering my design, I decided to use a table structure to display the library items instead of cards in a grid. A table will provide a more structured and semantically meaningful representation of the data compared to a grid. A table will also help address accessibility and UX concerns for mobile screens as it reduces the need for scrolling and resizing of a css grid to accommodate varying screen sizes as more data is added to the library.